Feature/privil mode spec update (#96)
* [rtl] Clear mprv on mret to non M-mode

This is specification change between the v1.11 and v1.12 privileged
architectures. Previously mprv wasn't altered on mret. Now if returning
to a privilege level other than M mode mprv must be cleared.

* [rtl, dv] Add new CSRs for latest priviledged spec

This adds the following CSRs to support the v1.12 priviledged spec.

 - MSTATUSH
 - MCONFIGPTR
 - MENVCFG
 - MENVCFGH

MCONFIGPTR is read only and has its value provided by a ibex_pkg
parameter CSR_MCONFIGPTR_VALUE which is set to 0. Implementors can alter
this value if needed.

All the other CSRs ignore writes and read as 0.

---------

Co-authored-by: Greg Chadwick <gac@lowrisc.org>